requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
asyncpg==0.30.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.3
cryptography==44.0.0
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.2.18
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.115.8
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.5
pillow==12.0.0
platformdirs==4.4.0
pluggy==1.5.0
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyGithub==2.5.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==311
pyzmq==27.0.2
requests==2.32.3
six==1.17.0
slack_sdk==3.34.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.45.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
